+++ examples/search/lib/app/search/domain/errors/erros.dart
@@ -1,9 +1,19 @@
-class Failure implements Exception {}
+class Failure implements Exception {
+  const Failure();
+}
 
-class InvalidSearchText extends Failure {}
+class InvalidSearchText extends Failure {
+  const InvalidSearchText();
+}
 
-class EmptyList extends Failure {}
+class EmptyList extends Failure {
+  const EmptyList();
+}
 
-class ErrorSearch extends Failure {}
+class ErrorSearch extends Failure {
+  const ErrorSearch();
+}
 
-class DatasourceResultNull extends Failure {}
+class DatasourceResultNull extends Failure {
+  const DatasourceResultNull();
+}