+++ triple/test/middleware_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: cascade_invocations
+
 import 'dart:async';
 
 import 'package:meta/meta.dart';
@@ -6,12 +8,15 @@ import 'package:triple/triple.dart';
 
 void main() {
   test('custom middleware', () {
-    final counter = Counter([]);
+    final counter = Counter(
+      const [],
+    );
     counter.increment();
     expect(counter.state, 3);
   });
 }
 
+// ignore: must_be_immutable
 class Counter extends TestImplements<Exception, int> {
   Counter(List<bool> list) : super(0, list);
 
@@ -27,7 +32,9 @@ class Counter extends TestImplements<Exception, int> {
   }
 }
 
-abstract class TestImplements<Error extends Object, State extends Object> extends Store<Error, State> {
+// ignore: must_be_immutable
+abstract class TestImplements<Error extends Object, State extends Object>
+    extends Store<Error, State> {
   final List<bool> list;
 
   TestImplements(State initialState, this.list) : super(initialState);