+++ triple_test/lib/triple_test.dart
@@ -1,0 +1,5 @@
+library triple_test;
+
+export 'src/store_mock.dart';
+export 'src/store_test.dart';
+export 'src/store_when.dart';