+++ triple/CHANGELOG.md
@@ -1,3 +1,31 @@
- 0.0.1
+  ## [1.5.0+1] - 2022-02-24
+  - Added `MementoMixin.clearHistory` method.
+  - Remove @protected of `MementoMixin.undo` and `MementoMixin.redo` methods.
+  ## [1.4.0] - 2021-10-21
+  - Added `HydratedMixin.hasInitiated` flag.
 
-- Initial version, created by Stagehand
+  ## [1.3.3] - 2021-10-21
+  - Added Store.when for a value of one of three mapped possibilities.
+  - Added @protected on update, setError, setLoading.
+
+  ## [1.3.0+1] - 2021-08-20
+  
+  - Added resolvers;
+  ## [1.2.0+3] - 2021-07-17
+  
+  - Added HydratedMixin and HydratedDelegate;
+
+  ## [1.1.0] - 2021-07-10
+
+ - Remove dartz dependency (Use **EitherAdapter**)
+ - Added EitherAdapter interface
+  ## [1.0.2] - 2021-05-10
+
+ - fix executeEither
+
+ ## [1.0.0] - 2021-03-03
+
+The Initial version providers:
+- abstract Stores
+- Triple Objects
+- Triple Tracking